# swagger.yaml
swagger: '2.0'
info:
  description: 'A REST API used to identify cross-references from requirements. Requirements can be in a database, or in an html document-html. Cross-references are dependencies explicitly stated, either internal (between requirements of the same project), or external (between a requirement and an external source).'
  version: '1.0'
  title: Cross-reference detection RESTful API
  contact:
    name: UPC-GESSI (OPENReq)
    url: 'http://openreq.eu/'
  license:
    name: EPL-v1.0
    url: 'https://www.eclipse.org/legal/epl-v10.html'
host: 'https://api.openreq.eu/cross-reference-detection'
basePath: /
...
paths:
...
  /upc/cross-reference-detection/reqPrefix:
    get:
      tags:
      - req-prefix-controller
      summary: Get prefixes
      description: Gets the list of prefixes of a given company. These prefixes are used to identify the requirements.
      operationId: getGrammarUsingGET
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - name: company
        in: query
        description: Company
        required: true
        type: string
      responses:
        '0':
          description: 'Non content: There is no content to submit.'
        '200':
          description: 'OK: The request has succeeded.'
          schema:
            type: string
        '201':
          description: 'Created: The request has been fulfilled and has resulted in one or more new resources being created.'
          schema:
            type: string
        '401':
          description: 'Unauthorized: The request has not been applied because it lacks valid authentication credentials for the target resource.'
        '403':
          description: 'Forbidden: The server understood the request but refuses to authorize it.'
        '404':
          description: 'Not Found: The server could not find what was requested by the client.'
        '500':
          description: Internal Server Error. For more information see ‘message’ in the Response Body.
    post:
      tags:
      - req-prefix-controller
      summary: Store prefixes
      description: 'Given a company, stores the list of prefixes used to identify requirements, which will be used in the grammar rules to extract requirement cross-references.'
      operationId: uploadGrammarUsingPOST
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - in: body
        name: prefixes
        description: The list of prefixes
        required: true
        schema:
          $ref: '#/definitions/PrefixList'
      - name: company
        in: query
        description: Company
        required: true
        type: string
      responses:
        '0':
          description: 'Non content: There is no content to submit.'
        '200':
          description: 'OK: The request has succeeded.'
          schema:
            type: string
        '201':
          description: 'Created: The request has been fulfilled and has resulted in one or more new resources being created.'
          schema:
            type: string
        '401':
          description: 'Unauthorized: The request has not been applied because it lacks valid authentication credentials for the target resource.'
        '403':
          description: 'Forbidden: The server understood the request but refuses to authorize it.'
        '404':
          description: 'Not Found: The server could not find what was requested by the client.'
        '500':
          description: Internal Server Error. For more information see ‘message’ in the Response Body.
    put:
      tags:
      - req-prefix-controller
      summary: Update prefixes
      description: 'Given a company, updates its existing list of prefixes used to identify requirements, which will be used in the grammar rules to extract requirement cross-references'
      operationId: updateGrammarUsingPUT
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - in: body
        name: prefixes
        description: The list of prefixes
        required: true
        schema:
          $ref: '#/definitions/PrefixList'
      - name: company
        in: query
        description: Company
        required: true
        type: string
      responses:
        '0':
          description: 'Non content: There is no content to submit.'
        '200':
          description: 'OK: The request has succeeded.'
          schema:
            type: string
        '201':
          description: 'Created: The request has been fulfilled and has resulted in one or more new resources being created.'
          schema:
            type: string
        '401':
          description: 'Unauthorized: The request has not been applied because it lacks valid authentication credentials for the target resource.'
        '403':
          description: 'Forbidden: The server understood the request but refuses to authorize it.'
        '404':
          description: 'Not Found: The server could not find what was requested by the client.'
        '500':
          description: Internal Server Error. For more information see ‘message’ in the Response Body.
    delete:
      tags:
      - req-prefix-controller
      summary: Delete prefixes
      description: Deletes the prefixes of a given company
      operationId: deleteGrammarUsingDELETE
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - name: company
        in: query
        description: Company
        required: true
        type: string
      responses:
        '0':
          description: 'Non content: There is no content to submit.'
        '200':
          description: 'OK: The request has succeeded.'
          schema:
            type: string
        '201':
          description: 'Created: The request has been fulfilled and has resulted in one or more new resources being created.'
          schema:
            type: string
        '204':
          description: No Content
        '401':
          description: 'Unauthorized: The request has not been applied because it lacks valid authentication credentials for the target resource.'
        '403':
          description: 'Forbidden: The server understood the request but refuses to authorize it.'
        '404':
          description: 'Not Found: The server could not find what was requested by the client.'
        '500':
          description: Internal Server Error. For more information see ‘message’ in the Response Body.
definitions:
  PrefixList:
    type: object
    properties:
      prefixes:
        type: array
        items:
          type: string
...